Working prototype of the COO grid
