+++ include/lvox/voxel/chunked_grid.hpp
@@ -2,8 +2,6 @@
 #define LVOX_VOXEL_CHUNKED_GRID_HPP
 
 #include <atomic>
-#include <execution>
-#include <filesystem>
 
 #include <lvox/types.hpp>
 #include <lvox/voxel/bounded_grid.hpp>
@@ -11,15 +9,9 @@
 namespace lvox
 {
 
-namespace algorithms::pad_estimators
-{
-struct BeerLambert;
-struct ContactFrequency;
-struct UnequalPathLengthBeerLambert;
-} // namespace algorithms::pad_estimators
-
 class ChunkedGrid
 {
+    friend class GridCOOView;
   public:
     ChunkedGrid(const Bounds& bounds, double cell_size, bool compute_variance);
     ChunkedGrid(ChunkedGrid&& other);
@@ -27,16 +19,6 @@ class ChunkedGrid
     auto register_hit(const Index3D& idx) -> void;
     auto add_length_and_count(const Index3D& voxel_idx, double length, bool is_hit) -> void;
     auto add_length_count_and_variance(const Index3D& idx, double length, bool is_hit) -> void;
-
-    auto compute_pad(algorithms::pad_estimators::BeerLambert) -> void;
-    auto compute_pad(algorithms::pad_estimators::ContactFrequency) -> void;
-    auto compute_pad(algorithms::pad_estimators::UnequalPathLengthBeerLambert) -> void;
-
-    auto export_as_coo_to_h5(
-        const std::string&           dataset_name,
-        const std::filesystem::path& filename,
-        bool                         include_all_data = false
-    ) const -> void;
 
     auto bounded_grid() const -> const BoundedGrid& { return m_bounded_grid; }
 
@@ -57,7 +39,6 @@ class ChunkedGrid
         std::vector<double>       m_lengths;
         std::vector<double>       m_hits_lengths;
         std::vector<double>       m_lengths_variance;
-        std::vector<double>       m_pad;
         std::mutex                m_write_access;
     };
 
@@ -71,31 +52,6 @@ class ChunkedGrid
     unsigned int             m_chunks_z;
     size_t                   m_chunk_count;
     std::vector<a_chunk_ptr> m_chunks;
-
-    template <typename PadFunc>
-    auto compute_pad_impl(PadFunc&& func) -> void
-    {
-        std::for_each(
-            std::execution::par_unseq, m_chunks.begin(), m_chunks.end(), [func](auto& a_chunk) {
-                auto chunk = a_chunk.load(std::memory_order_relaxed);
-
-                if (!chunk)
-                    return;
-
-                for (unsigned int i = 0; i < VoxelChunk::s_cell_count; ++i)
-                {
-                    // TODO: Make the treshold configurable
-                    if (chunk->m_counts[i] < 5)
-                    {
-                        chunk->m_pad[i] = 0;
-                        continue;
-                    }
-
-                    chunk->m_pad[i] = func(chunk, i);
-                }
-            }
-        );
-    }
 
     auto get_or_create_chunk(size_t chunk_idx) -> chunk_ptr;
     // Returns a pair of the chunk index in the index of the voxel in the chunk
